% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_po.R
\name{mi_po}
\alias{mi_po}
\title{Multiple Imputation for Two-Phase Ordinal Data}
\usage{
mi_po(
  formula,
  formula_exposure,
  data_R1,
  data_R0,
  ods_pi,
  n_candidates,
  n_impute = 50,
  verbose_TF = FALSE,
  srs_TF = FALSE
)
}
\arguments{
\item{formula}{The main model formula for the outcome.}

\item{formula_exposure}{The imputation model formula for the expensive exposure.}

\item{data_R1}{A data frame for subjects selected into Phase 2 (complete-data).}

\item{data_R0}{A data frame for subjects not selected into Phase 2 (missing expensive covariate).}

\item{ods_pi}{Sampling probabilities used for ascertainment correction.}

\item{n_candidates}{Number of grid points for expensive covariate in imputation.}

\item{n_impute}{Number of imputation datasets (default is 50).}

\item{verbose_TF}{Logical; if TRUE, prints progress every 25 imputations.}

\item{srs_TF}{Logical; set to TRUE if the design is Simple Random Sampling.}
}
\value{
A list containing the pooled coefficients and standard errors calculated via Rubin's Rules.
}
\description{
Implements the MI approach for ordinal outcomes in two-phase designs.
}
