% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal2phase-package.R
\docType{package}
\name{ordinal2phase-package}
\alias{ordinal2phase}
\alias{ordinal2phase-package}
\title{ordinal2phase: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
