% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_po.R
\name{smle_po}
\alias{smle_po}
\title{Sieve Maximum Likelihood Estimation}
\usage{
smle_po(
  formula,
  data_R1,
  data_R0,
  Bspline_R1,
  Bspline_R0,
  Xterm,
  se_calc = TRUE,
  theta_init = NULL,
  hnscale = 1,
  tol = 1e-04,
  maxiter = 1000,
  print_nsteps = 10,
  verbose = FALSE
)
}
\arguments{
\item{formula}{Model formula.}

\item{data_R1}{A data frame for subjects selected into Phase 2.}

\item{data_R0}{A data frame for subjects not selected into Phase 2.}

\item{Bspline_R1}{Spline basis for subjects selected into Phase 2.}

\item{Bspline_R0}{Spline basis for subjects not selected into Phase 2.}

\item{Xterm}{Name of the expensive covariate.}

\item{se_calc}{Logical, calculate standard errors?}

\item{theta_init}{Initial values for parameters.}

\item{hnscale}{Scale for the perturbation bandwidth.}

\item{tol}{Convergence tolerance.}

\item{maxiter}{Maximum EM iterations.}

\item{print_nsteps}{Frequency of printing progress.}

\item{verbose}{Logical, print detailed output?}
}
\description{
Implements the SMLE approach for two-phase sampling with ordinal outcomes.
}
